created first cavalier ability + edited marauder stuff

+++ src/main/java/io/github/Gabriel/ExpertiseStylePlugin/AbilitySystem/AbilityItemTemplate.java
@@ -54,11 +54,11 @@ public class AbilityItemTemplate {
         return -1;
     }
 
-    public static void putAllAbilitesOnCooldown(Player player, int cooldown) {
-        player.setCooldown(player.getInventory().getItem(0).getType(), 20 * cooldown);
-        player.setCooldown(player.getInventory().getItem(1).getType(), 20 * cooldown);
-        player.setCooldown(player.getInventory().getItem(2).getType(), 20 * cooldown);
-        player.setCooldown(player.getInventory().getItem(3).getType(), 20 * cooldown);
+    public static void putAllAbilitesOnCooldown(Player player, double cooldown) {
+        player.setCooldown(player.getInventory().getItem(0).getType(), (int) (20 * cooldown));
+        player.setCooldown(player.getInventory().getItem(1).getType(), (int) (20 * cooldown));
+        player.setCooldown(player.getInventory().getItem(2).getType(), (int) (20 * cooldown));
+        player.setCooldown(player.getInventory().getItem(3).getType(), (int) (20 * cooldown));
     }
 
     public static NamespacedKey getImmovableKey() {
